Fix monaco localization
